Update forumla for build changes.
With the Sqitch development currently in-progress, it will be possible to
reduce things down to a single formula, with different features enabled
via `--with-*` options, and all depdencies installed in the desintation
directory, rather than in the Perl lib directories. The work is ongoing,
but I wanted to push these changes as a checkpoint so I lose no work.