Add SnowflakeReq
This Requirement detects the presence of the Snowflake ODBC driver and
causes the installation to fail if it can't be found. It also checks
for SnowSQL and provides a `note` method telling the user about how to
install it or configure it once Sqitch has been installed. These notes
are output by the `post_install` method, which simply checks all
requirements for a `notes` method. I expect to use this pattern for
Firebird, Oracle, Vertica, and Exasol support, as well.