Use Postgres and MySQL clients instead of servers
Depend on the `libpq` and `mysql-client` formulas instead of
`postgresql` and `mysql`. This allows Sqitch to manage Postgres and
MySQL databases but does not force the user to install the servers to
support them.

Brought about by the warning recently added to Homebrew that deprecates
the `postgresql` dependency in favor of specific versions of Postgres.
Thanks to @ZhongRuoyu for the pointers in the
[discussion](https://github.com/orgs/Homebrew/discussions/4391) of this
issue and to @Alam080 for reporting it (resolves #56).