Remove all references to `--without-*` options
According to [this discussion], support for them was removed back in
2018. No one complained here until January 2024, so the assumption is
that they were not much used.

More to the point, as of 8167e40, the Postgres and MySQL server
dependencies were replaced with client-only dependencies, reducing the
need to prevent their installation. People still might want to build
with their own `libpq` or `mysql` client libraries, but they just need
to be sure they're found in their search paths (including library and
include) ahead of the Homebrew paths. In any event, since 8167e40 there
have been no more servers fired up (one hopes).

If it turns out people really do need a way to prevent the driver and
client dependencies from being installed, we can add our own `--without`
options. But as of right now, it doesn't seem necessary.

Resolves #62.

While at it, switch to fenced code blocks in the README.

  [this discussion]: https://github.com/orgs/Homebrew/discussions/5428